+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 * Fixed total sales is syned even if product is not managing stock
 * Fixed duplicator class PHP notice when product is being edited in quick mode
 * Fixed random behaviour for product type sync 
+* Fixed tax class are not synced
 
 ###0.25